Data Processing function
SuperMain function is a processing data function for CW-DOS
measurements, analyze the experimental data separately and is related
each measure with the experimental calibration. Is based in Main2
function for the anailysis and is requered the initial guess for
operation. Return a .Mat file with the paramters optimizated for the
Main2 function for each measurement along wiht standard deviation and
aditional information.